requirements.txt
requests
python-dotenv
google-generativeai
pytest
flake8
pandas
sqlalchemy
flask
flask-sqlalchemy